Fixes ShortReaderBehavior

+++ diana-api/src/main/java/org/apache/diana/api/reader/ShortReader.java
@@ -4,8 +4,8 @@ package org.apache.diana.api.reader;
 import org.apache.diana.api.ReaderField;
 
 /**
- * Class to reads and converts to {@link Double}, first it verify if is Double if yes return itself then verifies if is
- * {@link Number} and use {@link Number#doubleValue()} otherwise convert to {@link String} and then {@link Double}
+ * Class to reads and converts to {@link Short}, first it verify if is Double if yes return itself then verifies if is
+ * {@link Number} and use {@link Number#shortValue()} otherwise convert to {@link String} and then {@link Short}
  *
  * @author Otávio Santana
  */
@@ -19,13 +19,13 @@ public final class ShortReader implements ReaderField {
     @Override
     public <T> T read(Class<T> clazz, Object value) {
 
-        if (Double.class.isInstance(value)) {
+        if (Short.class.isInstance(value)) {
             return (T) value;
         }
         if (Number.class.isInstance(value)) {
-            return (T) Double.valueOf(Number.class.cast(value).doubleValue());
+            return (T) Short.valueOf(Number.class.cast(value).shortValue());
         } else {
-            return (T) Double.valueOf(value.toString());
+            return (T) Short.valueOf(value.toString());
         }
     }
 }
